==> R/00_colour_palettes.R <==
# Qualitative palette
thesis_col <- c(
    "#881177", "#aa3355", "#cc6666",
    "#ee9944", "#eedd00", "#99dd55",
    "#44dd88", "#22ccbb", "#00bbcc",
    "#0099cc", "#3366bb", "#663399"
)

# Sequential palette
thesis_col_green <- c(
    "#d9f0a3", "#addd8e", "#78c679",
    "#41ab5d", "#238443", "#005a32",
    "#00441b"
)

# Sequential palette
thesis_col_blue <- c(
    "#d0d1e6", "#a6bddb", "#74a9cf",
    "#3690c0", "#0570b0", "#034e7b",
    "#023858"
)

# Sequential palette
thesis_col_red <- c(
    "#fcbba1", "#fc9272", "#fb6a4a",
    "#ef3b2c", "#cb181d", "#67000d",
    "#67000d"
)

# Sequential palette
thesis_col_purple <- c(
    "#dadaeb", "#bcbddc", "#9e9ac8",
    "#807dba", "#6a51a3", "#4a1486",
    "#4a1486"
)

# Sequential palette
thesis_col_orange <- c(
    "#fdd49e", "#fdbb84", "#fc8d59",
    "#ef6548", "#d7301f", "#990000"
)

# Diverging palette
thesis_col_div <- c(
    "#d73027", "#f46d43", "#fdae61",
    "#fee08b", "#d9ef8b", "#a6d96a",
    "#66bd63", "#1a9850"
)
